refactor(types): group agent-domain types under common/types/agent
