+++ packages/desktop/src/common/types/agent/agentModes.ts
