+++ packages/desktop/src/common/types/agent/assistantTypes.ts
