+++ packages/desktop/src/common/types/agent/detectedAgent.ts
