+++ packages/desktop/src/common/types/agent/hub.ts
