+++ packages/desktop/src/renderer/pages/guid/components/AssistantSelectionArea.tsx
@@ -20,7 +20,7 @@ import { resolveAvatarImageSrc } from '@/renderer/pages/settings/AssistantSettin
 import { CUSTOM_AVATAR_IMAGE_MAP } from '../constants';
 import styles from '../index.module.css';
 import type { AvailableAgent, EffectiveAgentInfo } from '../types';
-import type { Assistant } from '@/common/types/assistantTypes';
+import type { Assistant } from '@/common/types/agent/assistantTypes';
 import { Message } from '@arco-design/web-react';
 import { Plus, Robot } from '@icon-park/react';
 import React, { useCallback, useLayoutEffect, useMemo } from 'react';
